Use of encoding and decoding
In today's society, security is crucial in a variety of applications. As a result, secure password storage in the database is required, and encoded copies of strings must be saved. Python's language contains a function called "encode()" that encodes texts using the provided encoding scheme. The language specifies a number of encoding techniques. The encode() function in Python encodes a text using the provided encoding. UTF-8 will be used if no encoding is given.